Updated all vcproj files to incorporate new directory structure. Reviewed all settings for each project and also increased warning level to level 3 and resolved all compiler warnings. There are still a couple trivial link warnings that will be resolved soon.

+++ src/ZoneServer/noiseutils.h
@@ -1211,7 +1211,7 @@ namespace noise
         /// Constructor.
         WriterTER ():
           m_pSourceNoiseMap (NULL),
-          m_metersPerPoint (DEFAULT_METERS_PER_POINT)
+          m_metersPerPoint (static_cast<float>(DEFAULT_METERS_PER_POINT))
         {
         }
